Add static chips and refactor code
