+++ packages/extension/src/view/devtools/components/tabContent/cookies/components/filterColumn/getFilters.ts
